Country YaML file and country selector CSS
countries.yml: contains countries acceptable to the service. It also
contains alternative spellings, abbreviations.

Based on user research and analysis of commonly made mistakes from the
business, we included common misspellings of countries.

country-selector.css: this is an extract (hopefully all?) of the CSS
used to style the type ahead when rendered.